+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 [Semantic Versioning](https://semver.org/)
 
+## [4.1.6] - 2023-05-15
+### Improvement
+- `RcsbSearchMetadata` includes rcsb schema metadata (e.g. `rcsb_search_context`, `rcsb_description`)
+- `RcsbSearchAttributeList` list of all search attribute objects
+
 ## [4.1.5] - 2023-05-11
 ### Improvement
 - `RcsbSearchAttributeType` new type for defining search attributes